+++ R/memo.R
@@ -39,7 +39,7 @@ memo <- function (f, allow.null=FALSE) {
   # TODO provide a way to supply cache arguments, for example the algo to use, max size, storage strategy, etc
 
   # get cache
-  f.cache <- cache()
+  f.cache <- storage.init()
   
   # create the memo function
   f.memo <- function (memo.force=FALSE, memo.dryrun=FALSE) {
@@ -54,10 +54,10 @@ memo <- function (f, allow.null=FALSE) {
     hash <- hash(fc)
 
     # if force or cached
-    if (!memo.force && cache.has(f.cache, hash)) {
+    if (!memo.force && storage.has(f.cache, hash)) {
 
       # false if dry run otherwise cached value
-      if (memo.dryrun) FALSE else cache.get(f.cache, hash)
+      if (memo.dryrun) FALSE else storage.get(f.cache, hash)
       
     } else {
 
@@ -68,7 +68,7 @@ memo <- function (f, allow.null=FALSE) {
       if (!is.null(result) || allow.null) {
         
         # cache the result
-        if (!memo.dryrun) cache.set(f.cache, hash, result)
+        if (!memo.dryrun) storage.set(f.cache, hash, result)
       }
       
       result